Adding sound effects

+++ app/src/main/java/com/example/mgp2023/AudioManager.java
@@ -26,7 +26,7 @@ public class AudioManager {
         Release();
     }
 
-    public void PlayAudio(int _id, int _volume)
+    public void PlayAudio(int _id, float _volume)
     {
         if (audioMap.containsKey(_id))
         {
